Version 0.1
Core functionality to read rekordbox XML files, analyse the tracks in the collection and attempt to relocate any missing files on disk, then rewrite the XML with the repaired tracks.  An option is included to choose between re-writing the whole XML file containing all tracks and playlists in order to do a full library rebuild, or only writing the repaired tracks to XML and any related playlists they were in, for manual drag-and-drop of only the repaired tracks back into the main collection and playlists.

During the analysis stage the tool also detects any files in MacOS which will exceed the file path limit of 255 characters, and which files exist on disk but not in rekordbox, e.g. which tracks in your iTunes collection haven't been imported to rekordbox yet.

At the end of the process a detailed report is written to disk for review.