Fix file canonical path read errors
If the canonical path of a file fails to be read for any reason, the error will be reported to the console and the process will continue uninterrupted.

This issue was found on Windows where the SoundCloud plugin had been installed in rekordbox. Tracks in the collection related to the plugin (and possibly others) are exported to the XML with a location looking like file://localhost/soundcloud:tracks:1149223021 , causing errors on Windows due to the colons (macOS is fine.)